move expectation into class
